requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
py==1.4.27
pytest==2.7.0
wsgiref==0.1.2
GeoPy==1.10.0
python-geohash==0.8.5
Flask-Cors==2.1.0